+++ app/helpers/plans_helper.rb
@@ -1,0 +1,2 @@
+module PlansHelper
+end
