+++ app/views/plans/index.html.erb
@@ -1,13 +1,13 @@
-<div>
-  <%= form_with model: @plan, url: calendars_path, class: 'form' do |f| %>
+
+  <%= form_with model: @plan, url: plans_path, class: 'form' do |f| %>
     <%= f.label :日付を選択, class: 'lavel' %>
     <%= f.date_field :date, class: 'date-select' %>
     <%= f.label :予定を入力, class: 'lavel' %>
     <%= f.text_field :plan, class: 'text-input' %>
     <%= f.submit '保存' %>
   <% end %>
-</div>
-<div>
+
+
   <div class=calendar>
     <% @week_days.each do |day| %>
       <div class='item'>
@@ -24,4 +24,3 @@
       </div>
     <% end %>
   </div>
-</div>
