+++ config/routes.rb
@@ -1,5 +1,4 @@
 Rails.application.routes.draw do
-  # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
-  root 'calendars#index'
-  resources :calendars
+  root 'plans#index'
+  resources :plans
 end
